moving Primer::Image to Primer::Alpha::Image
