+++ lib/tasks/docs.rake
@@ -33,7 +33,7 @@ namespace :docs do
       Primer::Beta::Button,
       Primer::Alpha::Layout,
       Primer::HellipButton,
-      Primer::Image,
+      Primer::Alpha::Image,
       Primer::LocalTime,
       Primer::OcticonSymbolsComponent,
       Primer::ImageCrop,
