+++ static/audited_at.json
@@ -7,6 +7,7 @@
   "Primer::Alpha::Dialog::Footer": "",
   "Primer::Alpha::Dialog::Header": "",
   "Primer::Alpha::HiddenTextExpander": "",
+  "Primer::Alpha::Image": "",
   "Primer::Alpha::Layout": "",
   "Primer::Alpha::Layout::Main": "",
   "Primer::Alpha::Layout::Sidebar": "",
