+++ static/constants.json
@@ -92,6 +92,8 @@
   },
   "Primer::Alpha::HiddenTextExpander": {
   },
+  "Primer::Alpha::Image": {
+  },
   "Primer::Alpha::Layout": {
     "FIRST_IN_SOURCE_DEFAULT": "sidebar",
     "FIRST_IN_SOURCE_OPTIONS": [
@@ -334,6 +336,7 @@
     "DEFAULT_PADDING": "default",
     "DEFAULT_ROW_SCHEME": "default",
     "Header": "Primer::Beta::BorderBox::Header",
+    "HeaderTest": "Primer::Beta::BorderBox::HeaderTest",
     "PADDING_MAPPINGS": {
       "default": "",
       "condensed": "Box--condensed",
@@ -546,7 +549,8 @@
   "Primer::DetailsComponent": {
   },
   "Primer::Dropdown": {
-    "Menu": "Primer::Dropdown::Menu"
+    "Menu": "Primer::Dropdown::Menu",
+    "MenuTest": "Primer::Dropdown::MenuTest"
   },
   "Primer::Dropdown::Menu": {
     "AS_DEFAULT": "default",
