+++ static/statuses.json
@@ -7,6 +7,7 @@
   "Primer::Alpha::Dialog::Footer": "alpha",
   "Primer::Alpha::Dialog::Header": "alpha",
   "Primer::Alpha::HiddenTextExpander": "alpha",
+  "Primer::Alpha::Image": "alpha",
   "Primer::Alpha::Layout": "alpha",
   "Primer::Alpha::Layout::Main": "alpha",
   "Primer::Alpha::Layout::Sidebar": "alpha",
@@ -59,7 +60,7 @@
   "Primer::HellipButton": "alpha",
   "Primer::HiddenTextExpander": "deprecated",
   "Primer::IconButton": "beta",
-  "Primer::Image": "alpha",
+  "Primer::Image": "deprecated",
   "Primer::ImageCrop": "alpha",
   "Primer::LabelComponent": "beta",
   "Primer::LayoutComponent": "alpha",
