+++ test/components/alpha/image_test.rb
@@ -2,18 +2,18 @@
 
 require "test_helper"
 
-class PrimerImageTest < Minitest::Test
+class PrimerAlphaImageTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_render_without_loading_attribute_by_default
-    render_inline(Primer::Image.new(src: "https://github.com/github.png", alt: "alt"))
+    render_inline(Primer::Alpha::Image.new(src: "https://github.com/github.png", alt: "alt"))
 
     assert_selector("img[src='https://github.com/github.png'][alt='alt']")
     refute_selector("[loading]")
   end
 
   def test_lazy_loading
-    render_inline(Primer::Image.new(src: "https://github.com/github.png", alt: "alt", lazy: true))
+    render_inline(Primer::Alpha::Image.new(src: "https://github.com/github.png", alt: "alt", lazy: true))
 
     assert_selector("img[src='https://github.com/github.png'][alt='alt'][loading='lazy'][decoding='async']")
   end
@@ -21,7 +21,7 @@ class PrimerImageTest < Minitest::Test
   def test_restricts_allowed_system_arguments
     with_raise_on_invalid_options(true) do
       error = assert_raises(ArgumentError) do
-        render_inline(Primer::Image.new(src: "https://github.com/github.png", alt: "alt", tag: :div))
+        render_inline(Primer::Alpha::Image.new(src: "https://github.com/github.png", alt: "alt", tag: :div))
       end
 
       assert_includes(error.message, "This component has a fixed tag.")
@@ -30,7 +30,7 @@ class PrimerImageTest < Minitest::Test
 
   def test_strips_denied_system_arguments
     with_raise_on_invalid_options(false) do
-      render_inline(Primer::Image.new(src: "https://github.com/github.png", alt: "alt", tag: :div))
+      render_inline(Primer::Alpha::Image.new(src: "https://github.com/github.png", alt: "alt", tag: :div))
     end
 
     refute_selector("div")
